requirements.txt
Flask==3.0.3
Jinja2==3.1.4
MarkupSafe==2.1.5
Werkzeug==3.0.3
itsdangerous==2.1.2

requests==2.32.3
flask-cors==4.0.1

# Use rapidfuzz instead of fuzzywuzzy/python-Levenshtein (no C-extensions needed)
rapidfuzz==3.14.1

# Use BeautifulSoup instead of htmldom
beautifulsoup4==4.12.3

gunicorn==21.2.0
gspread==6.1.4
google-auth==2.34.0
requests-cache==1.2.1
flask-limiter==3.8.0
redis==5.0.8

